test: add more unit mock service layer tests in project sub-module
